docs: callback should create heartbeat per subscription (#3936)
While trying to optimize the number of heartbeats is great, it had a
side effect that it would require some error prone logic on the subgraph
level to correctly group subscriptions per router (and it also
complicates the logic of cancelling sibling subscriptions which may be
running in separate threads). By requiring a heartbeat per subscription
we can keep the logic simple and make it easier to adopt. Further
optimizations to the protocol might be added at some later date.