check for idle connection in graceful shutdown
if a connection was opened to the router but never received the intial
request, ti would not be considered by hyper's graceful shutdown
feature, and so the router would wayt indefinitely for the connection to
close.

This tends to happen when a brower opens the explorer or prometheus page
provided by the router, because browsers eagerly open new connections,
in anticipation of future traffic, while those pages only need a single
request to the router.