exposing topologySpreadConstraints on the helm chart (#3892)
This allows users to set `topologySpreadConstraints` which can be used
to more evenly spread pods out over nodes/AZs/etc. and prevent uneven
utilization of federation pods.

Fixes #3891

<!-- start metadata -->
---

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR
is marked ready-for-review.

- [X] Changes are compatible[^1]
- [X] Documentation[^2] completed
- [X] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [X] Manual Tests

**Exceptions**
Tested helm chart template generation locally

*Note any exceptions here*
N/A
**Notes**
N/A

[^1]: It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding
Manual Tests, please document the manual testing (extensively) in the
Exceptions.

---------

Co-authored-by: Gary Pennington <gary@apollographql.com>
Co-authored-by: Gary Pennington <garypen@gmail.com>