docs: update JWT authentication intro (#3988)
This PR streamlines the introduction for the JWT authentication page so
that:
- Implementation details are moved to relevant sections (for most idPs,
users won't need to consider whether their idP use one of the signature
algorithms supported by the [jsonwebtoken Rust
library](https://crates.io/crates/jsonwebtoken)
- It only mentions once that this is an enterprise feature (and uses the
new shared component to do so)

This PR targets the main branch so that it can be published
expeditiously (ideally before Summit).

<!-- start metadata -->
---

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR
is marked ready-for-review.

- [ ] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding
Manual Tests, please document the manual testing (extensively) in the
Exceptions.

---------

Co-authored-by: Edward Huang <18322228+shorgi@users.noreply.github.com>