Fix defaulting in router config (#4010)
There are two types of serde defaulting:

* container
* field

Container level defaulting will use an instance of the default
implementation and take missing fields from it. Field level defaulting
uses either the default implementation or optionally user supplied
function to initialize missing fields.

When using field level defaulting it is essential that the Default
implementation of a struct exactly match the serde annotations.

A test checks to ensure that field level defaulting is not used in
conjunction with a Default implementation and gives guidance on
resolution.

*Description here*

Fixes #4000

<!-- start metadata -->
---

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR
is marked ready-for-review.

- [ ] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding
Manual Tests, please document the manual testing (extensively) in the
Exceptions.

Co-authored-by: bryn <bryn@apollographql.com>