Primarily use Rust validation errors when `graphql_validation_mode: both` (#3994)
This PR upgrades to a new apollo-compiler version that supports
serializing errors to JSON. `ApolloDiagnostic`s are turned into the JSON
representation ASAP (as diagnostics are not serializable themselves) and
threads the JSON errors through the different error types so they end up
in responses unchanged. With the better quality errors, it would now be
feasible to switch entirely to `graphql_validation_mode: new`.

This PR also changes the `both` validation mode so it primarily uses the
Rust errors. Previously, it would run Rust validation and JS validation,
compare, and then throw away the Rust errors. Now it's the other way
around. I would suggest that we keep running `both` in our infra so we
continue to check against the JS implementation, as we don't yet have
good results from fuzzing validation yet.

This PR has one minor user-facing issue that I definitely couldn't
address in time and that maybe we should solve separately, but it was
also true for the JS validation. Rust validation runs **after** some
query transformations, eg. defer labeling, which means that the location
information in rust validation errors may be wrong. Ideally, we would
run Rust validation as early as possible on the raw user input, but I
got a bit stuck trying to move things around. It will be easier with
compiler 1.0 and if we don't need to also compare the results to JS.

<!-- start metadata -->
---

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR
is marked ready-for-review.

- [ ] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding
Manual Tests, please document the manual testing (extensively) in the
Exceptions.

---------

Co-authored-by: Iryna Shestak <shestak.irina@gmail.com>
Co-authored-by: Bryn Cooke <BrynCooke@gmail.com>