Unify resource handling
The telemetry code had separate logic for constructing resources to pass to tracing and metrics exporters. This has now been unified and tests added.

As a consequence, the default service.name for the router in tracing has now been brought into line with the otel spec, where if unspecified the value will be `unknown_service`.