Update test-log crate, rename snapshot files accordingly
The previous version of this proc-macro crate would presumably
wrap tests in a function named `test_impl`, which would be picked up
by the insta crate when choosing a file name for snapshots.

It apparently doesn’t do that anymore, so rename snapshot files
to use the actual test function name.