Make all `apollo_router_http_requests_total` metrics match (#4065)
Metrics definitions must match, including descriptions.

Currently we have the same metric scattered across the codebase. This is
not ideal, but we are refactoring metrics callsites.
In the meantime, let's just make things match so that users don't see an
error in the logs.


I tries to pull stuff out to consts, but I don't think this is possible
without changing the macros, and we don't actually want people to have
multiple callsites for metrics. The golden rule is one metric one
callsite.

Fixes #4047

---------

Co-authored-by: bryn <bryn@apollographql.com>
Co-authored-by: Jesse Rosenberger <git@jro.cc>