Fix memory issues in the apollo metrics exporter
1. Bound channels should really be bound.
It transpires that the futures channel isn't "bound" as tight as we
would like in the face of sender clones. Replacing the futures channel
with a tokio channel removes the possiblity of memory expansion due to
a unresponsive receiver. Now, if the receiver isn't responsive enough,
the channel will start rejecting metrics.

2. Altering the priority of the ordering of select! processing with the
`biased` statement means the router will favour submitting metrics
reports to studio over reading more data and updating reports. This is
fine because we do want to send data promptly when a `tick` occurs.

3. Speeding up the submit_report() function reduces the possibilty of
memory bloat when the studio ingress stops accepting metrics. I've taken
out the retries since the router will retry on the next tick.