GraphOS authorization directives: policy directive (#3751)
replaces #3583

This introduces the `@policy` directive, defined as follows:

```graphql
scalar federation__Policy
directive @policy(policies: [[federation__Policy!]!]!) on
    OBJECT
  | FIELD_DEFINITION
  | INTERFACE
  | SCALAR
  | ENUM
```

It evaluates a list of policies, represented as arbitrary strings, that
are stored in the context as a map String -> Option<bool>. A coprocessor
or rhai script at the supergraph stage will look at this list of
policies, and indicate in the map which were successful. The router will
then accept fields for which all the required policies are successful.
The nested array in the `policies` argument works in the same way as
`@requiresScopes`: the inner array is an AND between policies, all of
them have to be successful at the same time, while the outer array is an
OR between the inner arrays, at least one of these inner arrays must be
successful.