Improve the robustness of our kubeconform processing
Sometimes a new release of kubernetes becomes available well in advance
of kubeconform support. Because we always use the latest patch release
of a kubernetes to validate our Helm files, we can end up with a broken
CI pipeline.

This changes the logic so that we always try all of the patch releases
for a minor release using the first one we find before moving on to the
next minor release. This is fine, since it's more important to keep our
CI pipeline working than to be verifying our Helm templates against the
very latest patch release of kubernetes.