Lower default GraphQL parser recursion limit to 500
The previous default of 4096 was too high.
I can reliable make the Router process (compiled in release mode)
abort with a stack overflow with ~2400 nested selection sets.

`deeply_nested.rs` demonstrates it.
It as "benchmark" because that’s an easy way to get access to a Router
compiled in release mode.
It is not run on CI because CI currently does not compile in release mode.

Note that the Router cannot handle response JSON data more nested
than 128 levels anyway:`serde_json::Deserializer` has its own
recursion limit, hard-coded to 128. It can be disabled
(with both run-time *and* compile-time opt-in) but not changed.

Our corpus of 5.3 million operations from customers has 8 entries
that use more recursion than the new default.

See also:

* https://github.com/apollographql/apollo-rs/pull/662
* https://github.com/apollographql/apollo-rs/pull/721