Improve the robustness of our kubeconform processing (#4178)
Sometimes a new release of kubernetes becomes available well in advance
of kubeconform support. Because we always use the latest patch release
of a kubernetes to validate our Helm files, we can end up with a broken
CI pipeline.

This changes the logic so that we always try all of the patch releases
for a minor release using the first one we find before moving on to the
next minor release. This is fine, since it's more important to keep our
CI pipeline working than to be verifying our Helm templates against the
very latest patch release of kubernetes.

<!-- start metadata -->
---

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR
is marked ready-for-review.

- [ ] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding
Manual Tests, please document the manual testing (extensively) in the
Exceptions.