Lower default GraphQL parser recursion limit to 500 (#4205)
The previous default of 4096 was too high. I can reliable make the
Router process (compiled in release mode) abort with a stack overflow
with ~2400 nested selection sets.

`deeply_nested.rs` demonstrates it. It as "benchmark" because that’s an
easy way to get access to a Router compiled in release mode. It is not
run on CI because CI currently does not compile in release mode.

Note that the Router cannot handle response JSON data more nested than
128 levels anyway:`serde_json::Deserializer` has its own recursion
limit, hard-coded to 128. It can be disabled (with both run-time *and*
compile-time opt-in) but not changed.

Our corpus of 5.3 million operations from customers has 8 entries that
use more recursion than the new default.

See also:

* https://github.com/apollographql/apollo-rs/pull/662
* https://github.com/apollographql/apollo-rs/pull/721

<!-- start metadata -->
---

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR
is marked ready-for-review.

- [x] Changes are compatible[^1]
- [x] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [x] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding
Manual Tests, please document the manual testing (extensively) in the
Exceptions.