Improve the secure deployability of our helm chart and docker image (#3971)
Many kubernetes environments impose security restrictions on containers.
For example:
 - Don't run as root
 - Can't bind to ports < 1024

Right now our container runs the router as `root` on port 80 and this
makes it harder to deploy in such secure environments.

This change moves our helm chart's default port from 80 to 4000 and
introduces a new user `router` in our docker container which we will use
to run our router process.

fixes: #3856 

---
**Impact**

1. Changing the default port in the helm chart from 80 to 4000 is an
innocuous change. I can't see how this would impact anyone.
2. Changing the default user from `root` to `router` will have an
impact. You will no longer be able to `exec` to the executing container
(kubernetes or docker) and install packages, etc... The container is now
"locked down", by default. Good for security, but less convenient for
support/debugging.

If you wish to restore the behaviour of the router to execute as `root`
and listen on port 80, that can be achieved with this configuration:

```
router:
  configuration:
    supergraph:
      listen: 0.0.0.0:80
securityContext:
  runAsUser: 0
```
I don't recommend that you do this for debugging/support (there are more
secure ways to achieve your goals in kubernetes), but am documenting for
completeness.

<!-- start metadata -->
---

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR
is marked ready-for-review.

- [ ] Changes are compatible[^1]
- [x] Documentation[^2] completed
- [x] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [x] Manual Tests

**Exceptions**

Manual testing of the impact of this change was performed on a local PC
(using docker) and in a Kubernetes cluster. The changes are working as
described. It's not viable to automate this testing.

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding
Manual Tests, please document the manual testing (extensively) in the
Exceptions.