use a custom service instead of private entries in the context
if we use the OneShotAsyncCheckpoint layer, then the only way to
communicate between the request and response size is to use the context.
We previously used the private entries, but when multiple subgraphs
requests are done in parallel, that same entry would be overwritten.
Instead of creating entries in context for each subgraph query, we make
a service that will keep the relevant data in its stack while the inner
service is called