docs(kubernetes/docker): Update relative path for configuration links (#4295)
*Description here*

Some of the links on the
[kubernetes](https://www.apollographql.com/docs/router/containerization/kubernetes/)
and
[docker](https://www.apollographql.com/docs/router/containerization/docker/)
pages refer to [a configuration
page](https://www.apollographql.com/docs/configuration/overview) two
levels up resulting in a 404

<img width="675" alt="image"
src="https://github.com/apollographql/router/assets/146107359/b3d1c260-4977-4153-b255-fe82744bbb9a">

Quickest way to find one is go to the Kubernetes page and click on
`command-line options and YAML configuration`

Fixes #**issue_number**

<!-- start metadata -->
---

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR
is marked ready-for-review.

- [ ] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding
Manual Tests, please document the manual testing (extensively) in the
Exceptions.