Disable running doctests
They are usually low value: demonstrating basic usage of API
that is already execised in parts of other tests.
However they are costly since each doctest is compiled as a separate
executable that depends on the apollo_router crate.
For reasons we don’t fully understand, linking takes a
pathologically long time and, especially on Linux, large amount of memory.
Many linker running in parallel can run out of memory. For example:

https://app.circleci.com/pipelines/github/apollographql/router/17749/workflows/5e2cd87a-c10b-41b8-900e-e5d517b8abe8/jobs/116288?invite=true#step-116-326592_69