Update v1.35.0 CHANGELOG.md with additional details around #4228 (#4329)
## Motivation

Updates Changelog that left-out "Coprocessor" impact/aspect of the fix,
which was as applicable as the Rhai aspect (same root cause). I've
already updated the original #4228 PR body to reflect this improved
scope.

## Markdown Preview Below
---

### Rhai scripts or coprocessors no longer prevent traces from appearing
in Apollo Studio ([PR
#4228](https://github.com/apollographql/router/pull/4228))

Previously, trace reports were not appearing in Apollo Studio's
Operations view when the Router was configured with either coprocessors
or Rhai script. That issue has been resolved in this release.

By [@bnjjj](https://github.com/bnjjj) in
https://github.com/apollographql/router/pull/4228