Make observable metrics clean up once they are dropped (#4353)
When observable metrics are created any callbacks are registered with
the metrics pipeline. They are never cleaned up. This means that we may
slowly leak memory if this feature is ever used.

Currently the metrics layer does NOT use this and instead calls
`observe` directly on gauges. This means that we don't currently leak,
but this shouldn't actually work as according to the rust docs for
`observe`:

```
Records the state of the instrument.
It is only valid to call this within a callback. If called outside of the registered callback it should have no effect on the instrument, and an error will be reported via the error handler.
```

Fixing this means that we get rid of the thread that is constantly
refreshing gauges between exports.

Fixes #4352

<!-- start metadata -->
---

**Checklist**

Complete the checklist (and note appropriate exceptions) before the PR
is marked ready-for-review.

- [ ] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]: It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]: Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]: Tick whichever testing boxes are applicable. If you are adding
Manual Tests, please document the manual testing (extensively) in the
Exceptions.

Co-authored-by: bryn <bryn@apollographql.com>